chore: move hospital common components
