+++ src/components/hospital/Badge.tsx
