+++ src/components/hospital/HospitalBasicInfo/styled.ts
