+++ src/components/hospital/HospitalDetail/index.tsx
@@ -3,7 +3,7 @@ import { ChevronLeftRounded } from "@/stories/Icon/ChevronLeft";
 import { useRouter } from "next/router";
 import PossibleAnimalList from "@components/common/hospital/PossibleAnimalList";
 import HospitalBasicInfo from "@components/common/hospital/HospitalBasicInfo";
-import Stats from "@components/common/hospital/Stats";
+import Stats from "@components/hospital/Stats";
 import Skeleton from "@/stories/common/Skeleton";
 import HospitalReview from "@components/hospital/HospitalReview/HospitalReview";
 import HospitalDetailReview from "@components/hospital/HospitalDetailReview";
