+++ src/components/hospital/PossibleAnimalList/styled.ts
