+++ src/components/hospital/Stats.tsx
